@@ -1,7 +1,7 @@
 ﻿using FluentScheduler;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Min.Web.BackgroundJobs;
 
@@ -11,21 +11,24 @@ namespace Min.Web
     {
         public static void Main(string[] args)
         {
-            var webHost = CreateWebHostBuilder(args).Build();
+            var host = CreateWebHostBuilder(args).Build();
 
-            JobManager.JobFactory = new JobFactory(webHost.Services);
+            JobManager.JobFactory = new JobFactory(host.Services);
             JobManager.Initialize(new FluentSchedulerRegistry());
             JobManager.JobException += info =>
             {
-                var logger = webHost.Services.GetRequiredService<ILogger<Program>>();
+                var logger = host.Services.GetRequiredService<ILogger<Program>>();
                 logger.LogError(info.Exception, "Unhandled exception in job");
             };
 
-            webHost.Run();
+            host.Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
+        public static IHostBuilder CreateWebHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                });
     }
 }
